Api endpoints now working (I hope so)

+++ rpc/main.py
@@ -2,9 +2,12 @@ from typing import Optional
 from pylon.core.tools import log
 from pylon.core.tools import web
 from pydantic import parse_obj_as, ValidationError
+from requests import HTTPError
+
 from ..models.prompts import Prompt, Example
 from tools import rpc_tools, db
-from ..models.pd.prompts_pd import PromptModel, ExampleModel
+from ..models.pd.prompts_pd import PromptModel, ExampleModel, PromptUpdateModel, \
+    ExampleUpdateModel
 
 
 class RPC:
@@ -13,43 +16,53 @@ class RPC:
     def prompts_get_all(self, project_id: int, **kwargs) -> list[dict]:
         with db.with_project_schema_session(project_id) as session:
             prompts = session.query(Prompt).all()
-            return [prompt.dict() for prompt in prompts]
+            return [prompt.to_json() for prompt in prompts]
 
     @web.rpc("prompts_get_by_id", "get_by_id")
-    def prompts_get_by_id(self, project_id: int, prompt_id: int, **kwargs) -> dict:
+    def prompts_get_by_id(self, project_id: int, prompt_id: int, **kwargs) -> dict | None:
         with db.with_project_schema_session(project_id) as session:
-            prompt = session.query(Prompt, Example).filter(
+            prompt = session.query(Prompt).filter(
                 Prompt.id == prompt_id,
-                Prompt.id == Example.prompt_id
+            ).one_or_none()
+            if not prompt:
+                return None
+            examples = session.query(Example).filter(
+                Example.prompt_id == prompt_id,
             ).all()
-            return prompt.dict()
+            result = prompt.to_json()
+            result['examples'] = [example.to_json() for example in examples]
+            return result
 
     @web.rpc(f'prompts_create', "create")
-    @rpc_tools.wrap_exceptions(ValidationError)
     def prompts_create(self, project_id: int, prompt: dict, **kwargs) -> dict:
         prompt = PromptModel.validate(prompt)
         with db.with_project_schema_session(project_id) as session:
             prompt = Prompt(**prompt.dict())
             session.add(prompt)
             session.commit()
-            return prompt.dict()
+            return prompt.to_json()
 
     @web.rpc(f'prompts_update', "update")
-    @rpc_tools.wrap_exceptions(ValidationError)
-    def prompts_update(self, project_id: int, prompt: dict, **kwargs) -> dict:
-        prompt = PromptModel.validate(prompt)
+    def prompts_update(self, project_id: int, prompt: dict, **kwargs) -> bool:
+        prompt = PromptUpdateModel.validate(prompt)
         with db.with_project_schema_session(project_id) as session:
-            prompt = session.query(Prompt).get(prompt['id'])
-            prompt.update(**prompt)
+            session.query(Prompt).filter(Prompt.id == prompt.id).update(
+                prompt.dict(exclude={'id'}, exclude_none=True)
+            )
             session.commit()
-            return prompt.dict()
+            updated_prompt = session.query(Prompt).get(prompt.id)
+            return updated_prompt.to_json()
 
     @web.rpc(f'prompts_delete', "delete")
-    @rpc_tools.wrap_exceptions(ValidationError)
     def prompts_delete(self, project_id: int, prompt_id: int, **kwargs) -> bool:
         with db.with_project_schema_session(project_id) as session:
             prompt = session.query(Prompt).get(prompt_id)
-            session.delete(prompt)
+            examples = session.query(Example).filter(Example.prompt_id == prompt_id).all()
+            if prompt:
+                session.delete(prompt)
+            for example in examples:
+                session.delete(example)
+
             session.commit()
             return True
 
@@ -59,34 +72,33 @@ class RPC:
     ) -> list[dict]:
         with db.with_project_schema_session(project_id) as session:
             examples = session.query(Example).filter(Example.prompt_id == prompt_id).all()
-            return [example.dict() for example in examples]
+            return [example.to_json() for example in examples]
 
     @web.rpc(f'prompts_create_example', "create_example")
-    @rpc_tools.wrap_exceptions(ValidationError)
     def prompts_create_example(self, project_id: int, example: dict, **kwargs) -> dict:
         example = ExampleModel.validate(example)
         with db.with_project_schema_session(project_id) as session:
-            session.query(Prompt).get_or_404(example['prompt_id'])
-            example = Example(**example)
+            example = Example(**example.dict())
             session.add(example)
             session.commit()
-            return example.dict()
+            return example.to_json()
 
     @web.rpc(f'prompts_update_example', "update_example")
-    @rpc_tools.wrap_exceptions(ValidationError)
-    def prompts_update_example(self, project_id: int, example: dict, **kwargs) -> dict:
-        example = ExampleModel.validate(example)
+    def prompts_update_example(self, project_id: int, example: dict, **kwargs) -> bool:
+        example = ExampleUpdateModel.validate(example)
         with db.with_project_schema_session(project_id) as session:
-            example = session.query(Example).get(example.id)
-            example.update(**example)
+            session.query(Example).filter(Example.id == example.id).update(
+                example.dict(exclude={'id'}, exclude_none=True)
+            )
             session.commit()
-            return example.dict()
+            updated_example = session.query(Example).get(example.id)
+            return updated_example.to_json()
 
     @web.rpc(f'prompts_delete_example', "delete_example")
-    @rpc_tools.wrap_exceptions(ValidationError)
     def prompts_delete_example(self, project_id: int, example_id: int, **kwargs) -> bool:
         with db.with_project_schema_session(project_id) as session:
             example = session.query(Example).get(example_id)
-            session.delete(example)
-            session.commit()
+            if example:
+                session.delete(example)
+                session.commit()
             return True
